Restructure types files #29. Update imports.
