+++ src/ts/championship.type.ts
