+++ src/ts/lastUpdate.type.ts
@@ -1,0 +1,4 @@
+export interface LastUpdate {
+    championship: string;
+    date: Date;
+}
